+++ library/src/test/java/org/mustangproject/ZUGFeRD/ZF2PushTest.java
@@ -119,8 +119,8 @@ public class ZF2PushTest extends TestCase {
 					 .load(SOURCE_PDF)) {
 
 			byte[] b = {12, 13};
-			ze.attachFile("one.pdf", b, "Application/PDF", "Alternative");
-			ze.attachFile("two.pdf", b, "Application/PDF", "Alternative");
+			ze.attachFile("one.pdf", b, "application/pdf", "Alternative");
+			ze.attachFile("two.pdf", b, "application/pdf", "Alternative");
 			ze.setTransaction(new Invoice().setDueDate(new Date()).setIssueDate(new Date()).setDeliveryDate(new Date()).setSender(new TradeParty(orgname, "teststr", "55232", "teststadt", "DE").addTaxID(taxID)).setOwnVATID("DE0815").setRecipient(new TradeParty("Franz Müller", "teststr.12", "55232", "Entenhausen", "DE").addVATID("DE4711").setContact(new Contact("Franz Müller", "01779999999", "franz@mueller.de", "teststr. 12", "55232", "Entenhausen", "DE"))).setNumber(number).addItem(new Item(new Product("Testprodukt", "", "C62", new BigDecimal(19)), price, new BigDecimal(1.0)))
 
 			);