+++ History.md
@@ -1,7 +1,7 @@
 2.0.2
-=====
-
-- #197 xml attachment errors
+=======
+- #197 file attachments
+- Charges/Allowances CategoryCode improvement PR #198 Thanks to weclapp-dev
 
 2.0.1
 =====