+++ History.md
@@ -1,5 +1,5 @@
 2.0.2
-=====
+=======
 - #197 file attachments
 - Charges/Allowances CategoryCode improvement PR #198 Thanks to weclapp-dev