+++ History.md
@@ -1,5 +1,7 @@
 2.0.2
 =======
+2020-11-25
+
 - #197 file attachments
 - Charges/Allowances CategoryCode improvement PR #198 Thanks to weclapp-dev
 - support reverse charge